Continue refactor
